Adding nodes and ties and their attributes

These functions allow users to add nodes, ties, or attributes to the nodes or ties of a network. The add_*() functions operate similarly to in {igraph}.

Not all functions have methods available for all object classes. Below are the currently implemented S3 methods:

igraph	network	tbl_graph

add_nodes	1	1	1

add_ties	1	1	1

delete_nodes	1	0	0

add_nodes(.data, nodes, attribute = NULL)

delete_nodes(.data, nodes)

add_ties(.data, ties, attribute = NULL)

add_node_attribute(.data, attr_name, vector)

add_tie_attribute(.data, attr_name, vector)

Arguments

.data

An object of a {manynet}-consistent class:

matrix (adjacency or incidence) from {base} R

edgelist, a data frame from {base} R or tibble from {tibble}

igraph, from the {igraph} package

network, from the {network} package

tbl_graph, from the {tidygraph} package

nodes

The number of nodes to be added.

attribute

A named list to be added as tie or node attributes.

ties

The number of ties to be added or an even list of ties.

attr_name

Name of the new attribute in the resulting object.

vector

A vector of values for the new attribute.

Value

A data object of the same class as the function was given.

Functions

add_nodes(): Add additional nodes to a network

delete_nodes(): Delete nodes in a network

add_ties(): Add additional ties to a network

add_node_attribute(): Add a vector of values to a network as a nodal attribute.

add_tie_attribute(): Add a vector of values to a network as a tie attribute.

See also

Examples

  other <- create_filled(4) %>% mutate(name = c("A", "B", "C", "D"))
  add_nodes(other, 4, list(name = c("Matthew", "Mark", "Luke", "Tim")))
#> # A labelled, undirected network with 8 nodes and 6 ties
#> # A tibble: 8 × 1
#>   name   
#>   <chr>  
#> 1 A      
#> 2 B      
#> 3 C      
#> 4 D      
#> 5 Matthew
#> 6 Mark   
#> # ℹ 2 more rows
#> # A tibble: 6 × 2
#>    from    to
#>   <int> <int>
#> 1     1     2
#> 2     1     3
#> 3     1     4
#> 4     2     3
#> 5     2     4
#> 6     3     4
  add_tie_attribute(other, "weight", c(1, 2, 2, 2, 1, 2))
#> # A labelled, weighted, undirected network with 4 nodes and 6 ties
#> # A tibble: 4 × 1
#>   name 
#>   <chr>
#> 1 A    
#> 2 B    
#> 3 C    
#> 4 D    
#> # A tibble: 6 × 3
#>    from    to weight
#>   <int> <int>  <dbl>
#> 1     1     2      1
#> 2     1     3      2
#> 3     1     4      2
#> 4     2     3      2
#> 5     2     4      1
#> 6     3     4      2