Describing attributes of nodes or ties in a network

These functions extract certain attributes from given network data. They are also useful as helpers within other functions.

node_names(.data)

node_mode(.data)

node_attribute(.data, attribute)

tie_attribute(.data, attribute)

tie_weights(.data)

tie_signs(.data)

Arguments

.data

An object of a {manynet}-consistent class:

matrix (adjacency or incidence) from {base} R

edgelist, a data frame from {base} R or tibble from {tibble}

igraph, from the {igraph} package

network, from the {network} package

tbl_graph, from the {tidygraph} package

attribute

Character string naming an attribute in the object.

Value

node_*() and tie_*() always return vectors the same length as the number of nodes or ties in the network, respectively.

Functions

node_names(): Extracts the names of the nodes in a network.

node_mode(): Extracts the mode of the nodes in a network.

node_attribute(): Extracts an attribute's values for the nodes in a network.

tie_attribute(): Extracts an attribute's values for the edges in a network.

tie_weights(): Extracts the weights of the edges in a network.

tie_signs(): Extracts the signs of the edges in a network.

See also

Examples

node_names(ison_southern_women)
#>  [1] "EVELYN"    "LAURA"     "THERESA"   "BRENDA"    "CHARLOTTE" "FRANCES"  
#>  [7] "ELEANOR"   "PEARL"     "RUTH"      "VERNE"     "MYRA"      "KATHERINE"
#> [13] "SYLVIA"    "NORA"      "HELEN"     "DOROTHY"   "OLIVIA"    "FLORA"    
#> [19] "E1"        "E2"        "E3"        "E4"        "E5"        "E6"       
#> [25] "E7"        "E8"        "E9"        "E10"       "E11"       "E12"      
#> [31] "E13"       "E14"      
node_mode(ison_southern_women)
#>    EVELYN     LAURA   THERESA    BRENDA CHARLOTTE   FRANCES   ELEANOR     PEARL 
#>     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE 
#>      RUTH     VERNE      MYRA KATHERINE    SYLVIA      NORA     HELEN   DOROTHY 
#>     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE 
#>    OLIVIA     FLORA        E1        E2        E3        E4        E5        E6 
#>     FALSE     FALSE      TRUE      TRUE      TRUE      TRUE      TRUE      TRUE 
#>        E7        E8        E9       E10       E11       E12       E13       E14 
#>      TRUE      TRUE      TRUE      TRUE      TRUE      TRUE      TRUE      TRUE 
#> attr(,"class")
#> [1] "node_mark" "logical"  
#> attr(,"mode")
#>    EVELYN     LAURA   THERESA    BRENDA CHARLOTTE   FRANCES   ELEANOR     PEARL 
#>     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE 
#>      RUTH     VERNE      MYRA KATHERINE    SYLVIA      NORA     HELEN   DOROTHY 
#>     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE 
#>    OLIVIA     FLORA        E1        E2        E3        E4        E5        E6 
#>     FALSE     FALSE      TRUE      TRUE      TRUE      TRUE      TRUE      TRUE 
#>        E7        E8        E9       E10       E11       E12       E13       E14 
#>      TRUE      TRUE      TRUE      TRUE      TRUE      TRUE      TRUE      TRUE 
#> attr(,"class")
#> [1] "node_mark" "logical"  
node_attribute(ison_lotr, "Race")
#>  [1] "Human"  "Human"  "Hobbit" "Elf"    "Human"  "Elf"    "Elf"    "Human" 
#>  [9] "Human"  "Hobbit" "Elf"    "Maiar"  "Dwarf"  "Elf"    "Hobbit" "Hobbit"
#> [17] "Hobbit" "Hobbit" "Maiar"  "Human"  "Maiar"  "Elf"    "Ent"    "Human" 
#> [25] "Maiar"  "Human"  "Elf"    "Human"  "Elf"    "Human"  "Elf"    "Elf"   
#> [33] "Hobbit" "Human"  "Maiar"  "Human" 
tie_attribute(ison_algebra, "task_tie")
#> NULL
tie_weights(to_mode1(ison_southern_women))
#>   [1] 6 6 7 3 4 3 3 3 2 2 2 2 2 1 2 1 1 6 6 3 4 4 3 2 2 2 2 2 1 1 1 6 4 4 4 4 3
#>  [38] 3 3 3 2 2 2 2 1 1 4 4 4 3 2 2 2 2 2 1 1 1 2 2 2 1 1 1 1 3 2 2 1 1 1 1 1 1
#>  [75] 1 3 2 2 2 2 2 1 1 1 2 2 2 2 2 2 1 2 1 1 3 3 2 2 2 2 2 1 1 4 3 3 3 3 3 1 1
#> [112] 4 4 3 4 3 1 1 6 5 4 5 1 1 6 6 4 1 1 6 3 2 2 3 1 1 1 1 2
tie_signs(ison_marvel_relationships)
#>   [1] -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1  1  1  1  1 -1 -1 -1  1 -1 -1 -1 -1 -1 -1
#>  [26] -1 -1 -1 -1 -1 -1 -1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#>  [51]  1  1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1  1  1  1  1  1  1  1  1  1  1  1  1  1
#>  [76]  1  1  1  1  1  1  1  1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1  1  1  1  1  1  1  1
#> [101]  1  1  1  1  1 -1 -1 -1 -1 -1 -1  1  1  1  1  1  1  1  1 -1 -1 -1 -1  1  1
#> [126]  1  1  1  1  1  1  1  1  1  1  1  1  1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1  1
#> [151]  1  1  1  1  1  1  1  1  1  1 -1 -1 -1 -1 -1 -1  1  1  1  1  1  1  1  1  1
#> [176]  1  1  1  1  1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1  1  1  1  1  1  1  1  1
#> [201]  1 -1 -1 -1 -1 -1 -1  1  1  1  1  1  1  1  1  1 -1 -1 -1 -1 -1 -1 -1 -1  1
#> [226]  1  1  1  1  1  1  1  1  1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1
#> [251] -1  1 -1 -1 -1 -1 -1 -1 -1 -1 -1  1  1  1  1  1  1  1  1  1  1  1 -1 -1 -1
#> [276] -1  1  1 -1  1  1  1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1  1  1 -1 -1
#> [301]  1 -1 -1 -1 -1 -1 -1 -1  1  1  1  1  1  1 -1  1  1  1  1  1  1  1  1  1 -1
#> [326] -1 -1 -1 -1 -1 -1 -1  1  1  1  1  1  1  1  1  1 -1 -1 -1 -1 -1 -1 -1 -1 -1
#> [351] -1 -1 -1 -1 -1 -1 -1  1  1  1  1  1  1  1  1  1  1  1  1 -1 -1 -1 -1 -1 -1
#> [376] -1 -1 -1  1  1  1  1  1  1  1  1  1  1  1  1 -1 -1 -1 -1 -1 -1 -1 -1 -1  1
#> [401]  1  1  1  1  1  1  1  1  1 -1 -1 -1 -1 -1 -1 -1  1  1 -1  1  1 -1 -1 -1 -1
#> [426] -1 -1 -1 -1 -1 -1 -1 -1  1  1 -1 -1  1  1  1  1  1  1 -1 -1 -1  1 -1 -1 -1
#> [451] -1 -1 -1  1  1 -1 -1 -1 -1 -1  1  1  1  1  1  1  1 -1 -1 -1 -1 -1 -1 -1 -1
#> [476] -1 -1 -1  1  1  1  1  1 -1 -1 -1 -1 -1  1  1  1  1  1 -1 -1  1  1  1  1  1
#> [501]  1  1 -1  1  1  1  1 -1  1  1  1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1 -1  1  1  1
#> [526] -1 -1 -1 -1  1  1  1  1 -1 -1  1  1  1  1  1 -1 -1 -1  1  1  1 -1 -1 -1 -1
#> [551]  1  1 -1 -1  1 -1  1 -1