Two-mode network of Love Actually characters and their scene appearances (Robinson 2015)

Love Actually is a 2003 British romantic comedy film. David Robinson's data is a two-mode network of 20 characters and their appearances across 76 scenes, as parsed from the script.

Added to this are multiplex, one-mode networks of character relationships, including "romantic", "family", "friendship", and "professional" ties. These were added by Korakot Janteerasakul from the following source: https://en.wikipedia.org/wiki/Love_Actually#/media/File:Love_Actually_(2003)_Interconnections.svg.

data(fict_actually)

Format

#> -- # Love actually interactions ------------------------------------------------
#> # A labelled, multiplex, two-mode network of 20 characters and 76 scenes and
#> 162 appearance, 3 family, 7 friendship, 10 professional, and 7 romance ties
#>
#> -- Nodes
#> # A tibble: 96 x 6
#>   type  name    Actor          Gender Nationality Occupation
#>   <lgl> <chr>   <chr>          <chr>  <chr>       <chr>
#> 1 FALSE Aurelia Lúcia Moniz    Female Portuguese  Assistant
#> 2 FALSE Billy   Bill Nighy     Male   British     Singer
#> 3 FALSE Colin   Kris Marshall  Male   British     Travels to America
#> 4 FALSE Daniel  Liam Neeson    Male   British     Widower
#> 5 FALSE Harry   Alan Rickman   Male   British     Agency Director
#> 6 FALSE John    Martin Freeman Male   British     Nude stand-in
#> # i 90 more rows
#>
#> -- Ties
#> # A tibble: 189 x 3
#>    from    to type
#>   <int> <int> <chr>
#> 1     1     7 romance
#> 2     1    51 appearance
#> 3     1    57 appearance
#> 4     1    58 appearance
#> 5     1    63 appearance
#> 6     1    95 appearance
#> # i 183 more rows
#>

References

Robinson, David. 2015. "Analyzing networks of characters in 'Love Actually'". http://varianceexplained.org/r/love-actually-network/

Author

David Robinson