Tools for joining lists of networks, graphs, and matrices

These functions offer tools for joining lists of manynet-consistent objects (matrices, igraph, tidygraph, or network objects). Joining expects a list of objects and returns a single network object.

from_subgraphs(.data)

from_egos(.data)

from_waves(.data)

from_slices(.data, remove.duplicates = FALSE)

Arguments

.data

An object of a manynet-consistent class:

matrix (adjacency or incidence) from {base} R

edgelist, a data frame from {base} R or tibble from {tibble}

igraph, from the {igraph} package

network, from the {network} package

tbl_graph, from the {tidygraph} package

remove.duplicates

Should duplicates be removed? By default FALSE. If TRUE, duplicated edges are removed.

Functions

from_subgraphs(): Returns a single network object from a list of subgraphs.

from_egos(): Returns a single network object from a list of egos.

from_waves(): Returns a single network object from a list of waves.

from_slices(): Returns a single network object from a list of slices.

See also

Examples

ison_adolescents %>%
  mutate(unicorn = sample(c("yes", "no"), 8, replace = TRUE)) %>%
  to_subgraphs(attribute = "unicorn") %>%
  from_subgraphs()
#> Joining with `by = join_by(name, unicorn)`
#> # A labelled, undirected tbl_graph with 8 nodes and 6 ties
#> # A tibble: 8 × 2
#>   name  unicorn
#>   <chr> <chr>  
#> 1 Betty yes    
#> 2 Sue   yes    
#> 3 Dale  yes    
#> 4 Alice no     
#> 5 Jane  no     
#> 6 Pam   no     
#> # ℹ 2 more rows
#> # A tibble: 6 × 2
#>    from    to
#>   <int> <int>
#> 1     1     2
#> 2     2     3
#> 3     4     5
#> 4     4     6
#> 5     6     7
#> 6     7     8
ison_adolescents %>%
  activate(edges) %>%
  to_egos() %>%
  from_egos()
#> # A labelled, directed tbl_graph with 8 nodes and 10 arcs
#> # A tibble: 8 × 1
#>   name 
#>   <chr>
#> 1 Betty
#> 2 Sue  
#> 3 Alice
#> 4 Jane 
#> 5 Pam  
#> 6 Carol
#> # ℹ 2 more rows
#> # A tibble: 10 × 2
#>    from    to
#>   <int> <int>
#> 1     1     2
#> 2     2     3
#> 3     2     7
#> 4     3     7
#> 5     2     5
#> 6     3     5
#> # ℹ 4 more rows
ison_adolescents %>%
  mutate_ties(wave = sample(1:4, 10, replace = TRUE)) %>%
  to_waves(attribute = "wave") %>%
  from_waves()
#> # A longitudinal, labelled, directed tbl_graph with 8 nodes and 10 arcs
#> # A tibble: 8 × 1
#>   name 
#>   <chr>
#> 1 Betty
#> 2 Alice
#> 3 Jane 
#> 4 Carol
#> 5 Sue  
#> 6 Pam  
#> # ℹ 2 more rows
#> # A tibble: 10 × 3
#>    from    to  wave
#>   <int> <int> <int>
#> 1     1     5     1
#> 2     2     7     1
#> 3     3     7     1
#> 4     2     6     1
#> 5     4     8     1
#> 6     5     2     4
#> # ℹ 4 more rows
ison_adolescents %>%
  mutate_ties(time = 1:10, increment = 1) %>% 
  add_ties(c(1,2), list(time = 3, increment = -1)) %>% 
  to_slices(slice = c(5,7)) %>%
  from_slices()
#> # A labelled, multiplex, weighted, directed tbl_graph with 5 nodes and 10 arcs
#> # A tibble: 5 × 1
#>   name 
#>   <chr>
#> 1 Sue  
#> 2 Alice
#> 3 Jane 
#> 4 Dale 
#> 5 Pam  
#> # A tibble: 10 × 3
#>    from    to weight
#>   <int> <int>  <dbl>
#> 1     1     2      1
#> 2     2     3      1
#> 3     1     4      1
#> 4     2     4      1
#> 5     1     2      1
#> 6     2     3      1
#> # ℹ 4 more rows