Marking networks classes

These functions implement logical tests for networks' classes.

is_manynet() marks a network TRUE if it is compatible with {manynet} functions.

is_edgelist() marks a network TRUE if it is an edgelist.

is_graph() marks a network TRUE if it contains graph-level information.

is_list() marks a network TRUE if it is a (non-igraph) list of networks, for example a set of ego networks or a dynamic or longitudinal set of networks.

is_longitudinal() marks a network TRUE if it contains longitudinal, panel data.

is_dynamic() marks a network TRUE if it contains dynamic, time-stamped data

All is_*() functions return a logical scalar (TRUE or FALSE).

is_manynet(.data)

is_graph(.data)

is_edgelist(.data)

is_list(.data)

is_longitudinal(.data)

is_dynamic(.data)

Arguments

.data

An object of a manynet-consistent class:

matrix (adjacency or incidence) from {base} R

edgelist, a data frame from {base} R or tibble from {tibble}

igraph, from the {igraph} package

network, from the {network} package

tbl_graph, from the {tidygraph} package

Value

TRUE if the condition is met, or FALSE otherwise.

See also

Examples

is_manynet(create_filled(2))
#> [1] TRUE
is_graph(create_star(2))
#> [1] TRUE
is_edgelist(matrix(c(2,2), 1, 2))
#> [1] FALSE
is_edgelist(as_edgelist(matrix(c(2,2), 1, 2)))
#> [1] TRUE
is_longitudinal(create_tree(5, 3))
#> [1] FALSE
is_dynamic(create_tree(3))
#> [1] FALSE