One-mode multiplex, directed network of managers of a high-tech company (Krackhardt 1987)

21 managers of a company of just over 100 employees manufactured high-tech equipment on the west coast of the United States. Three types of ties were collected:

friends: managers' answers to the question "Who is your friend?"

advice: managers' answers to the question "To whom do you go to for advice?"

reports: "To whom do you report?" based on company reports

The data is anonymised, but four nodal attributes are included:

age: the manager's age in years

tenure: the manager's length of service

level: the manager's level in the corporate hierarchy, where 3 = CEO, 2 = Vice President, and 1 = manager

dept: one of four departments, B, C, D, E, with the CEO alone in A

data(ison_hightech)

Format

#> # A multiplex, directed network with 21 nodes and 312 arcs
#> # A tibble: 21 x 4
#>     age tenure level dept
#>   <dbl>  <dbl> <dbl> <chr>
#> 1    33      9     1 E
#> 2    42     20     2 E
#> 3    40     13     1 C
#> 4    33      8     1 E
#> 5    32      3     1 C
#> 6    59     28     1 B
#> # i 15 more rows
#> # A tibble: 312 x 3
#>    from    to type
#>   <int> <int> <chr>
#> 1     1     2 friends
#> 2     1     2 advice
#> 3     1     2 reports
#> 4     1     4 friends
#> 5     1     4 advice
#> 6     1     8 friends
#> # i 306 more rows

References

Krackhardt, David. 1987. "Cognitive social structures". Social Networks 9: 104-134.