One-mode multiplex, longitudinal network of a Zambian tailor shop (Kapferer 1972)

Bruce Kapferer observed interactions in a tailor shop in Zambia, then Northern Rhodesia, over a period of ten months. His interest was in the changing patterns of alliance among the 39 workers during extended negotiations for higher wages.

Two layers of tie were recorded, each at two times seven months apart:

instrumental: work- and assistance-related interactions

sociational: friendship and other socioemotional interactions

The data are of particular interest because an abortive strike occurred after the first set of observations, and a successful strike took place after the second, so the two waves bracket a change in the workers' collective capacity.

data(ison_tailorshop)

Format

#> -- # Kapferer's Tailor Shop ----------------------------------------------------
#> # A longitudinal, labelled, multiplex, directed network of 39 workers and 256
#> instrumental arcs and 381 sociational ties over 2 waves
#>
#> -- Nodes
#> # A tibble: 39 x 1
#>   label
#>   <chr>
#> 1 Kamwefu
#> 2 Nkumbula
#> 3 Abraham
#> 4 Seams
#> 5 Chipata
#> 6 Donald
#> # i 33 more rows
#>
#> -- Ties
#> # A tibble: 637 x 4
#>    from    to layer         time
#>   <int> <int> <chr>        <int>
#> 1     3     1 instrumental     1
#> 2     4     1 instrumental     1
#> 3    14     1 instrumental     1
#> 4    11     2 instrumental     1
#> 5    12     2 instrumental     1
#> 6     1     3 instrumental     1
#> # i 631 more rows
#>

Details

Worker names have been title-cased. While the network as a whole is listed as directed, only the instrumental layer is directed. The sociational layer is symmetric. The info component records the directedness of each layer. Coercion to another class reciprocates that layer again, because 'igraph' or 'network' objects can only be directed or undirected.

References

Kapferer, Bruce. 1972. Strategy and Transaction in an African Factory. Manchester: Manchester University Press.

Dosdall, Henrik. 2019. "Kapferer (1972): Strategy and Transaction in an African Factory". In Schlüsselwerke der Netzwerkforschung, 249-252. Wiesbaden: Springer. doi:10.1007/978-3-658-21742-6_66