Making networks of package dependencies

Researchers regularly need to work with a variety of external data formats. The following functions offer ways to import from some common external file formats into objects that {manynet} and other graph/network packages in R can work with:

read_matrix() imports adjacency matrices from Excel/csv files.

read_edgelist() imports edgelists from Excel/csv files.

read_nodelist() imports nodelists from Excel/csv files.

read_pajek() imports Pajek (.net or .paj) files.

read_ucinet() imports UCINET files from the header (.##h).

read_dynetml() imports DyNetML interchange format for rich social network data.

read_graphml() imports GraphML files.

read_cran(pkg = "all")

Source

https://www.r-bloggers.com/2016/01/r-graph-objects-igraph-vs-network/

Arguments

pkg

The name

Details

Note that these functions are not as actively maintained as others in the package, so please let us know if any are not currently working for you or if there are missing import routines by raising an issue on Github.

See also

Examples

# mnet <- read_cran()
# mnet <- to_ego(mnet, "manynet", max_dist = 2)
# graphr(mnet, layout = "hierarchy", 
#        edge_color = "type", node_color = "Compilation")