Layout algorithms based on configurational positions

Configurational layouts locate nodes at symmetric coordinates to help illustrate particular configurations. Currently "triad" and "quad" layouts are available. The "configuration" layout will choose the appropriate configurational layout automatically.

layout_tbl_graph_configuration(.data, circular = FALSE, times = 1000)

layout_tbl_graph_dyad(.data, circular = FALSE, times = 1000)

layout_tbl_graph_triad(.data, circular = FALSE, times = 1000)

layout_tbl_graph_tetrad(.data, circular = FALSE, times = 1000)

layout_tbl_graph_pentad(.data, circular = FALSE, times = 1000)

layout_tbl_graph_hexad(.data, circular = FALSE, times = 1000)

Arguments

.data

An object of a manynet-consistent class:

matrix (adjacency or incidence) from {base} R

edgelist, a data frame from {base} R or tibble from {tibble}

igraph, from the {igraph} package

network, from the {network} package

tbl_graph, from the {tidygraph} package

circular

Should the layout be transformed into a radial representation. Only possible for some layouts. Defaults to FALSE.

times

Maximum number of iterations, where appropriate

See also