Marking nodes for selection based on measures

These functions return logical vectors the length of the nodes in a network identifying which hold certain properties or positions in the network.

node_is_random() marks one or more nodes at random.

node_is_max() and node_is_min() are more generally useful for converting the results from some node measure into a mark-class object. They can be particularly useful for highlighting which node or nodes are key because they minimise or, more often, maximise some measure.

node_is_random(.data, size = 1)

node_is_max(node_measure, ranks = 1)

node_is_min(node_measure, ranks = 1)

node_is_mean(node_measure, ranks = 1)

Arguments

.data

An object of a manynet-consistent class:

matrix (adjacency or incidence) from {base} R

edgelist, a data frame from {base} R or tibble from {tibble}

igraph, from the {igraph} package

network, from the {network} package

tbl_graph, from the {tidygraph} package

size

The number of nodes to select (as TRUE).

node_measure

An object created by a node_ measure.

ranks

The number of ranks of max or min to return. For example, ranks = 3 will return TRUE for nodes with scores equal to any of the top (or, for node_is_min(), bottom) three scores. By default, ranks = 1.

See also

Examples

node_is_random(ison_brandes, 2)
#>   V1    V2    V3    V4    V5    V6    V7    V8    V9    V10   V11  
#> 1 FALSE FALSE FALSE FALSE TRUE  FALSE FALSE FALSE FALSE FALSE TRUE 
#node_is_max(migraph::node_degree(ison_brandes))
#node_is_min(migraph::node_degree(ison_brandes))
#node_is_mean(node_degree(ison_brandes))