Marking ties based on structural properties

These functions return logical vectors the length of the ties in a network identifying which hold certain properties or positions in the network.

tie_is_multiple() marks ties that are multiples.

tie_is_loop() marks ties that are loops.

tie_is_reciprocated() marks ties that are mutual/reciprocated.

tie_is_feedback() marks ties that are feedback arcs causing the network to not be acyclic.

tie_is_bridge() marks ties that cut or act as articulation points in a network.

tie_is_path() marks ties on a path from one node to another.

They are most useful in highlighting parts of the network that are particularly well- or poorly-connected.

tie_is_multiple(.data)

tie_is_loop(.data)

tie_is_reciprocated(.data)

tie_is_feedback(.data)

tie_is_bridge(.data)

tie_is_path(.data, from, to, all_paths = FALSE)

Arguments

.data

An object of a manynet-consistent class:

matrix (adjacency or incidence) from {base} R

edgelist, a data frame from {base} R or tibble from {tibble}

igraph, from the {igraph} package

network, from the {network} package

tbl_graph, from the {tidygraph} package

from

The index or name of the node from which the path should be traced.

to

The index or name of the node to which the path should be traced.

all_paths

Whether to return a list of paths or sample just one. By default FALSE, sampling just a single path.

See also

Examples

tie_is_multiple(ison_marvel_relationships)
#>   `Abomination-Beast` `Abomination-Colossus` `Abomination-Cyclops`
#> 1 FALSE               FALSE                  FALSE                
#> # ... with 555 more values from this nodeset unprinted. Use `print(..., n = Inf)` to print all values.
tie_is_loop(ison_marvel_relationships)
#>   `Abomination-Beast` `Abomination-Colossus` `Abomination-Cyclops`
#> 1 FALSE               FALSE                  FALSE                
#> # ... with 555 more values from this nodeset unprinted. Use `print(..., n = Inf)` to print all values.
tie_is_reciprocated(ison_algebra)
#>   `1->5` `1->5` `1->8` `1->9` `1->9` `1->10` `1->10` `1->11` `1->12` `1->12`
#> 1 TRUE   TRUE   FALSE  TRUE   TRUE   TRUE    TRUE    TRUE    TRUE    TRUE   
#> # ... with 269 more values from this nodeset unprinted. Use `print(..., n = Inf)` to print all values.
tie_is_feedback(ison_algebra)
#>   `1->5` `1->5` `1->8` `1->9` `1->9` `1->10` `1->10` `1->11` `1->12` `1->12`
#> 1 TRUE   TRUE   TRUE   TRUE   TRUE   TRUE    TRUE    TRUE    TRUE    TRUE   
#> # ... with 269 more values from this nodeset unprinted. Use `print(..., n = Inf)` to print all values.
tie_is_bridge(ison_brandes)
#>   `1-3` `2-3` `3-4` `4-5` `4-6` `5-7` `6-7` `6-8` `7-9` `8-9` `9-10` `9-11`
#> 1 TRUE  TRUE  TRUE  FALSE FALSE FALSE FALSE FALSE FALSE FALSE TRUE   TRUE  
ison_adolescents %>%
  mutate_ties(route = tie_is_path(from = "Jane", to = 7)) 
#> 
#> ── # The Adolescent Society ────────────────────────────────────────────────────
#> # A labelled, multiplex, undirected network of 8 adolescents and 10 friendships
#> ties
#> 
#> ── Nodes 
#> # A tibble: 8 × 1
#>   name 
#>   <chr>
#> 1 Betty
#> 2 Sue  
#> 3 Alice
#> 4 Jane 
#> 5 Dale 
#> 6 Pam  
#> # ℹ 2 more rows
#> 
#> ── Ties 
#> # A tibble: 10 × 3
#>    from    to route     
#>   <int> <int> <tie_mark>
#> 1     1     2 FALSE     
#> 2     2     3 FALSE     
#> 3     3     4 TRUE      
#> 4     2     5 FALSE     
#> 5     3     5 FALSE     
#> 6     4     5 FALSE     
#> # ℹ 4 more rows
#> 
  #graphr(edge_colour = "route")