Describing network properties

These functions extract certain attributes from given network data. They are also useful as helpers within other functions.

network_nodes(.data)

network_ties(.data)

network_dims(.data)

network_node_attributes(.data)

network_tie_attributes(.data)

Arguments

.data

An object of a {manynet}-consistent class:

matrix (adjacency or incidence) from {base} R

edgelist, a data frame from {base} R or tibble from {tibble}

igraph, from the {igraph} package

network, from the {network} package

tbl_graph, from the {tidygraph} package

Value

network_*() functions always relate to the overall graph or network, usually returning a scalar. network_dims() returns an integer of the number of nodes in a one-mode network, or two integers representing the number of nodes in each nodeset in the case of a two-mode network. network_*_attributes() returns a string vector with the names of all node or tie attributes in the network.

Functions

network_nodes(): Returns the total number of nodes (of any mode) in a network.

network_ties(): Returns the number of edges in a network.

network_dims(): Returns the dimensions of a network in a vector as long as the number of modes in the network.

network_node_attributes(): Returns a vector of nodal attributes in a network

network_tie_attributes(): Returns a vector of edge attributes in a network

See also

Examples

network_nodes(ison_southern_women)
#> [1] 32
network_ties(ison_southern_women)
#> [1] 93
network_dims(ison_southern_women)
#> [1] 18 14
network_dims(to_mode1(ison_southern_women))
#> [1] 18
  network_node_attributes(ison_lotr)
#> [1] "name" "Race"
  network_tie_attributes(ison_algebra)
#> [1] "friends" "social"  "tasks"