Tools for reformatting networks, graphs, and matrices

These functions offer tools for reformatting migraph-consistent objects (matrices, igraph, tidygraph, or network objects). Unlike the as_*() group of functions, these functions always return the same object type as they are given, only transforming these objects' properties.

Not all functions have methods available for all object classes. Below are the currently implemented S3 methods:

data.frame	igraph	matrix	network	tbl_graph

to_acyclic	0	1	1	1	1

to_directed	1	1	1	1	1

to_multilevel	0	1	1	0	1

to_named	1	1	1	1	1

to_onemode	0	1	1	0	1

to_reciprocated	1	1	1	1	1

to_redirected	1	1	1	1	1

to_simplex	0	1	1	0	1

to_twomode	0	1	0	1	1

to_undirected	1	1	1	1	1

to_uniplex	1	1	1	1	1

to_unnamed	1	1	1	1	1

to_unsigned	1	1	1	1	1

to_unweighted	1	1	1	1	1

to_uniplex(.data, edge)

to_undirected(.data)

to_directed(.data)

to_redirected(.data)

to_reciprocated(.data)

to_acyclic(.data)

to_unweighted(.data, threshold = 1)

to_unsigned(.data, keep = c("positive", "negative"))

to_unnamed(.data)

to_named(.data, names = NULL)

to_simplex(.data)

to_onemode(.data)

to_multilevel(.data)

to_twomode(.data, mark)

Arguments

.data

An object of a {manynet}-consistent class:

matrix (adjacency or incidence) from {base} R

edgelist, a data frame from {base} R or tibble from {tibble}

igraph, from the {igraph} package

network, from the {network} package

tbl_graph, from the {tidygraph} package

edge

Character string naming an edge attribute to retain from a graph.

threshold

For a matrix, the threshold to binarise/dichotomise at.

keep

In the case of a signed network, whether to retain the "positive" or "negative" ties.

names

Character vector of the node names. NULL by default.

mark

A logical vector marking two types or modes. By default "type".

Value

All to_ functions return an object of the same class as that provided. So passing it an igraph object will return an igraph object and passing it a network object will return a network object, with certain modifications as outlined for each function.

Functions

to_uniplex(): Returns an object that includes only a single type of tie

to_undirected(): Returns an object that has any edge direction removed, so that any pair of nodes with at least one directed edge will be connected by an undirected edge in the new network. This is equivalent to the "collapse" mode in {igraph}.

to_directed(): Returns a directed object. Note that ties' direction will be randomly assigned. To flip the direction, use to_redirected(). To match the direction, use to_reciprocated().

to_redirected(): Returns an object that has any edge direction transposed, or flipped, so that senders become receivers and receivers become senders. This essentially has no effect on undirected networks or reciprocated ties.

to_reciprocated(): Returns an object where all ties are reciprocated.

to_acyclic(): Returns an object where all ties are acyclic.

to_unweighted(): Returns an object that has all edge weights removed.

to_unsigned(): Returns a network with either just the "positive" ties or just the "negative" ties

to_unnamed(): Returns an object with all vertex names removed

to_named(): Returns an object that has random vertex names added

to_simplex(): Returns an object that has all loops or self-ties removed

to_onemode(): Returns an object that has any type/mode attributes removed, but otherwise includes all the same nodes and ties. Note that this is not the same as to_mode1() or to_mode2(), which return only some of the nodes and new ties established by coincidence.

to_multilevel(): Returns a network that is not divided into two mode types but embeds two or more modes into a multimodal network structure.

to_twomode(): Returns a network that divides the nodes into two mode types.

See also

Examples

as_tidygraph(create_filled(5)) %>%
  mutate_ties(type = sample(1:2, 10, replace = TRUE)) %>%
  to_uniplex("type")
#> # A weighted, undirected tbl_graph with 5 nodes and 10 ties
#> # A tibble: 10 × 3
#>    from    to weight
#>   <int> <int>  <int>
#> 1     1     2      2
#> 2     1     3      2
#> 3     1     4      2
#> 4     1     5      2
#> 5     2     3      2
#> 6     2     4      2
#> # ℹ 4 more rows