Tools for transforming networks, graphs, and matrices

These functions offer tools for transforming migraph-consistent objects (matrices, igraph, tidygraph, or network objects). Transforming means that the returned object may have different dimensions than the original object.

Not all functions have methods available for all object classes. Below are the currently implemented S3 methods:

data.frame	igraph	list	matrix	network	tbl_graph

to_anti	1	1	0	1	1	1

to_blocks	1	1	0	1	1	1

to_eulerian	0	1	0	0	0	1

to_giant	1	1	0	1	1	1

to_matching	1	1	0	1	1	1

to_mode1	1	1	0	1	1	1

to_mode2	1	1	0	1	1	1

to_no_isolates	1	1	1	1	1	1

to_subgraph	1	1	0	1	1	1

to_ties	1	1	0	1	1	1

to_mode1(.data, similarity = c("count", "jaccard", "rand", "pearson", "yule"))

to_mode2(.data, similarity = c("count", "jaccard", "rand", "pearson", "yule"))

to_giant(.data)

to_subgraph(.data, ...)

to_ties(.data)

to_blocks(.data, membership, FUN = mean)

to_matching(.data, mark = "type")

to_eulerian(.data)

to_anti(.data)

to_no_isolates(.data)

Arguments

.data

An object of a {manynet}-consistent class:

matrix (adjacency or incidence) from {base} R

edgelist, a data frame from {base} R or tibble from {tibble}

igraph, from the {igraph} package

network, from the {network} package

tbl_graph, from the {tidygraph} package

similarity

Method for establishing ties, currently "count" (default), "jaccard", or "rand". "count" calculates the number of coinciding ties, and can be interpreted as indicating the degree of opportunities between nodes. "jaccard" uses this count as the numerator in a proportion, where the denominator consists of any cell where either node has a tie. It can be interpreted as opportunity weighted by participation. "rand", or the Simple Matching Coefficient, is a proportion where the numerator consists of the count of cells where both nodes are present or both are absent, over all possible cells. It can be interpreted as the (weighted) degree of behavioral mirroring between two nodes. "pearson" (Pearson's coefficient) and "yule" (Yule's Q) produce correlations for valued and binary data, respectively. Note that Yule's Q has a straightforward interpretation related to the odds ratio.

...

Arguments passed on to dplyr::filter

membership

A vector of partition memberships.

FUN

A function for summarising block content. By default mean. Other recommended options include median, sum, min or max.

mark

A logical vector marking two types or modes. By default "type".

Value

All to_ functions return an object of the same class as that provided. So passing it an igraph object will return an igraph object and passing it a network object will return a network object, with certain modifications as outlined for each function.

Functions

to_mode1(): Results in a weighted one-mode object that retains the row nodes from a two-mode object, and weights the ties between them on the basis of their joint ties to nodes in the second mode (columns)

to_mode2(): Results in a weighted one-mode object that retains the column nodes from a two-mode object, and weights the ties between them on the basis of their joint ties to nodes in the first mode (rows).

to_giant(): Returns an object that includes only the main component without any smaller components or isolates

to_subgraph(): Returns a network subgraph filtered on the basis of some node-related logical statement.

to_ties(): Returns a matrix (named if possible) where the edges are the nodes

to_blocks(): Returns a reduced graph from a given partition membership vector. Reduced graphs provide summary representations of network structures by collapsing groups of connected nodes into single nodes while preserving the topology of the original structures.

to_matching(): Returns a network with only matching ties

to_eulerian(): Returns a network with only the Eulerian path

to_anti(): Returns the complement of a network where only ties not present in the original network are included in the new network.

to_no_isolates(): Removes all nodes without ties

to_matching

to_matching() uses {igraph}'s max_bipartite_match() to return a network in which each node is only tied to one of its previous ties. The number of these ties left is its cardinality, and the algorithm seeks to maximise this such that, where possible, each node will be associated with just one node in the other mode or some other mark. The algorithm used is the push-relabel algorithm with greedy initialization and a global relabelling after every \(\frac{n}{2}\) steps, where \(n\) is the number of nodes in the network.

References

Goldberg, A V; Tarjan, R E (1986). "A new approach to the maximum flow problem". Proceedings of the eighteenth annual ACM symposium on Theory of computing – STOC '86. p. 136. doi:10.1145/12130.12144

See also

Examples

to_mode1(ison_southern_women)
#> # A labelled, weighted, undirected tbl_graph with 18 nodes and 139 ties
#> # A tibble: 18 × 1
#>   name     
#>   <chr>    
#> 1 EVELYN   
#> 2 LAURA    
#> 3 THERESA  
#> 4 BRENDA   
#> 5 CHARLOTTE
#> 6 FRANCES  
#> # ℹ 12 more rows
#> # A tibble: 139 × 3
#>    from    to weight
#>   <int> <int>  <dbl>
#> 1     1     2      6
#> 2     1     4      6
#> 3     1     3      7
#> 4     1     5      3
#> 5     1     6      4
#> 6     1     7      3
#> # ℹ 133 more rows
to_mode2(ison_southern_women)
#> # A labelled, weighted, undirected tbl_graph with 14 nodes and 66 ties
#> # A tibble: 14 × 1
#>   name 
#>   <chr>
#> 1 E1   
#> 2 E2   
#> 3 E3   
#> 4 E4   
#> 5 E5   
#> 6 E6   
#> # ℹ 8 more rows
#> # A tibble: 66 × 3
#>    from    to weight
#>   <int> <int>  <dbl>
#> 1     1     2      2
#> 2     1     3      3
#> 3     1     4      2
#> 4     1     5      3
#> 5     1     6      3
#> 6     1     8      3
#> # ℹ 60 more rows
#autographr(to_mode1(ison_southern_women))
#autographr(to_mode2(ison_southern_women))
to_ties(ison_adolescents)
#> # A labelled, undirected tbl_graph with 10 nodes and 20 ties
#> # A tibble: 10 × 1
#>   name      
#>   <chr>     
#> 1 Betty-Sue 
#> 2 Sue-Alice 
#> 3 Alice-Jane
#> 4 Sue-Dale  
#> 5 Alice-Dale
#> 6 Jane-Dale 
#> # ℹ 4 more rows
#> # A tibble: 20 × 2
#>    from    to
#>   <int> <int>
#> 1     1     2
#> 2     2     3
#> 3     1     4
#> 4     2     4
#> 5     4     5
#> 6     2     5
#> # ℹ 14 more rows
#autographr(to_ties(ison_adolescents))
to_matching(ison_southern_women)
#> # A labelled, two-mode tbl_graph with 32 nodes and 14 ties
#> # A tibble: 32 × 2
#>   name      type 
#>   <chr>     <lgl>
#> 1 EVELYN    FALSE
#> 2 LAURA     FALSE
#> 3 THERESA   FALSE
#> 4 BRENDA    FALSE
#> 5 CHARLOTTE FALSE
#> 6 FRANCES   FALSE
#> # ℹ 26 more rows
#> # A tibble: 14 × 2
#>    from    to
#>   <int> <int>
#> 1     1    19
#> 2     2    20
#> 3     3    21
#> 4     4    22
#> 5     5    23
#> 6     6    24
#> # ℹ 8 more rows
#autographr(to_matching(ison_southern_women))
  to_eulerian(delete_nodes(ison_konigsberg, "Lomse"))
#> IGRAPH 59217d5 DN-- 3 4 -- 
#> + attr: name (v/c)
#> + edges from 59217d5 (vertex names):
#> [1] Altstadt->Kneiphof Kneiphof->Vorstadt Vorstadt->Kneiphof Kneiphof->Altstadt
  #autographr(to_eulerian(delete_nodes(ison_konigsberg, "Lomse")))
to_anti(ison_southern_women)
#> # A labelled, two-mode tbl_graph with 32 nodes and 159 ties
#> # A tibble: 32 × 2
#>   type  name     
#>   <lgl> <chr>    
#> 1 FALSE EVELYN   
#> 2 FALSE LAURA    
#> 3 FALSE THERESA  
#> 4 FALSE BRENDA   
#> 5 FALSE CHARLOTTE
#> 6 FALSE FRANCES  
#> # ℹ 26 more rows
#> # A tibble: 159 × 2
#>    from    to
#>   <int> <int>
#> 1     1    25
#> 2     1    28
#> 3     1    29
#> 4     1    30
#> 5     1    31
#> 6     1    32
#> # ℹ 153 more rows
#autographr(to_anti(ison_southern_women))
ison_adolescents %>%
  activate(edges) %>%
  mutate(wave = sample(1995:1998, 10, replace = TRUE)) %>%
  to_waves(attribute = "wave") %>%
  to_no_isolates()
#> $`1998`
#> # A longitudinal, labelled, undirected tbl_graph with 2 nodes and 1 ties
#> # A tibble: 2 × 1
#>   name 
#>   <chr>
#> 1 Betty
#> 2 Sue  
#> # A tibble: 1 × 3
#>    from    to  wave
#>   <int> <int> <int>
#> 1     1     2  1998
#> 
#> $`1995`
#> # A longitudinal, labelled, undirected tbl_graph with 4 nodes and 3 ties
#> # A tibble: 4 × 1
#>   name 
#>   <chr>
#> 1 Sue  
#> 2 Alice
#> 3 Pam  
#> 4 Carol
#> # A tibble: 3 × 3
#>    from    to  wave
#>   <int> <int> <int>
#> 1     1     2  1995
#> 2     2     3  1995
#> 3     3     4  1995
#> 
#> $`1997`
#> # A longitudinal, labelled, undirected tbl_graph with 6 nodes and 3 ties
#> # A tibble: 6 × 1
#>   name 
#>   <chr>
#> 1 Sue  
#> 2 Alice
#> 3 Jane 
#> 4 Dale 
#> 5 Carol
#> 6 Tina 
#> # A tibble: 3 × 3
#>    from    to  wave
#>   <int> <int> <int>
#> 1     2     3  1997
#> 2     1     4  1997
#> 3     5     6  1997
#> 
#> $`1996`
#> # A longitudinal, labelled, undirected tbl_graph with 5 nodes and 3 ties
#> # A tibble: 5 × 1
#>   name 
#>   <chr>
#> 1 Sue  
#> 2 Alice
#> 3 Jane 
#> 4 Dale 
#> 5 Pam  
#> # A tibble: 3 × 3
#>    from    to  wave
#>   <int> <int> <int>
#> 1     2     4  1996
#> 2     3     4  1996
#> 3     1     5  1996
#>