Obtain overview of available network data

This function makes it easy to get an overview of available data:

table_data() returns a tibble with details of the network datasets included in the packages.

table_data(pkg = c("manynet", "migraph"), ...)

Arguments

pkg

String, name of the package.

...

Network marks, e.g. directed, twomode, or signed, that are used to filter the results.

Examples

table_data()
#> # A tibble: 22 × 13
#>    dataset            nodes  ties nattr tattr directed weighted twomode labelled
#>    <char>             <dbl> <dbl> <dbl> <dbl> <lgl>    <lgl>    <lgl>   <lgl>   
#>  1 ison_koenigsberg       4     7     3     1 FALSE    FALSE    FALSE   TRUE    
#>  2 irps_wwi               6    20     1     3 FALSE    FALSE    FALSE   TRUE    
#>  3 ison_adolescents       8    10     1     0 FALSE    FALSE    FALSE   TRUE    
#>  4 ison_brandes          11    12     1     0 FALSE    FALSE    FALSE   FALSE   
#>  5 ison_algebra          16   279     0     2 TRUE     TRUE     FALSE   FALSE   
#>  6 ison_monks            18   463     3     4 TRUE     TRUE     FALSE   TRUE    
#>  7 ison_hightech         21   312     4     1 TRUE     FALSE    FALSE   FALSE   
#>  8 ison_networkers       32   440     3     1 TRUE     TRUE     FALSE   TRUE    
#>  9 ison_southern_wom…    32    89     4     1 FALSE    FALSE    TRUE    TRUE    
#> 10 ison_karateka         34    78     2     1 FALSE    TRUE     FALSE   TRUE    
#> # ℹ 12 more rows
#> # ℹ 4 more variables: signed <lgl>, multiplex <lgl>, acyclic <lgl>,
#> #   attributed <lgl>
# to obtain list of all e.g. directed networks:
table_data(pkg = "manynet", directed)
#> # A tibble: 8 × 13
#>   dataset         nodes  ties nattr tattr directed weighted twomode labelled
#>   <char>          <dbl> <dbl> <dbl> <dbl> <lgl>    <lgl>    <lgl>   <lgl>   
#> 1 ison_algebra       16   279     0     2 TRUE     TRUE     FALSE   FALSE   
#> 2 ison_monks         18   463     3     4 TRUE     TRUE     FALSE   TRUE    
#> 3 ison_hightech      21   312     4     1 TRUE     FALSE    FALSE   FALSE   
#> 4 ison_networkers    32   440     3     1 TRUE     TRUE     FALSE   TRUE    
#> 5 fict_potter        64   544     4     1 TRUE     FALSE    FALSE   TRUE    
#> 6 ison_lawfirm       71  2571     7     1 TRUE     FALSE    FALSE   FALSE   
#> 7 fict_thrones      208   404    10     1 TRUE     FALSE    FALSE   TRUE    
#> 8 irps_blogs       1490 19090     3     0 TRUE     FALSE    FALSE   TRUE    
#> # ℹ 4 more variables: signed <lgl>, multiplex <lgl>, acyclic <lgl>,
#> #   attributed <lgl>
# to obtain overview of unique datasets:
table_data() %>% 
  dplyr::distinct(directed, weighted, twomode, signed, 
                 .keep_all = TRUE)
#> # A tibble: 7 × 13
#>   dataset             nodes  ties nattr tattr directed weighted twomode labelled
#>   <char>              <dbl> <dbl> <dbl> <dbl> <lgl>    <lgl>    <lgl>   <lgl>   
#> 1 ison_koenigsberg        4     7     3     1 FALSE    FALSE    FALSE   TRUE    
#> 2 irps_wwi                6    20     1     3 FALSE    FALSE    FALSE   TRUE    
#> 3 ison_algebra           16   279     0     2 TRUE     TRUE     FALSE   FALSE   
#> 4 ison_monks             18   463     3     4 TRUE     TRUE     FALSE   TRUE    
#> 5 ison_hightech          21   312     4     1 TRUE     FALSE    FALSE   FALSE   
#> 6 ison_southern_women    32    89     4     1 FALSE    FALSE    TRUE    TRUE    
#> 7 ison_karateka          34    78     2     1 FALSE    TRUE     FALSE   TRUE    
#> # ℹ 4 more variables: signed <lgl>, multiplex <lgl>, acyclic <lgl>,
#> #   attributed <lgl>