Functions that have been renamed, superseded, or are no longer working

Generally these functions have been superseded or renamed. Upon using them, a message is provided directing the user to the new function. However, at this stage of package development, we generally clear older defunct functions at each minor release, and so you are strongly encouraged to use the new functions/names/syntax wherever possible and update your scripts accordingly.

pkg_data(pkg = "manynet")

node_mode(.data)

autographr(
  .data,
  layout,
  labels = TRUE,
  node_color,
  node_shape,
  node_size,
  node_group,
  edge_color,
  edge_size,
  ...
)

autographs(netlist, waves, based_on = c("first", "last", "both"), ...)

autographd(
  tlist,
  layout,
  labels = TRUE,
  node_color,
  node_shape,
  node_size,
  edge_color,
  edge_size,
  keep_isolates = TRUE,
  ...
)

node_optimal(.data)

node_kernighanlin(.data)

node_edge_betweenness(.data)

node_fast_greedy(.data)

node_leading_eigen(.data)

node_walktrap(.data)

node_infomap(.data)

node_spinglass(.data)

node_fluid(.data)

node_leiden(.data)

node_louvain(.data)

node_roulette(.data)

node_components(.data)

node_weak_components(.data)

node_strong_components(.data)

node_equivalence(.data)

node_structural_equivalence(.data)

node_regular_equivalence(.data)

node_automorphic_equivalence(.data)

network_transmissibility(diff_model)

network_infection_length(diff_model)

network_reproduction(diff_model)

network_immunity(diff_model, normalized = TRUE)

network_hazard(diff_model)

network_density(.data)

network_components(.data)

network_cohesion(.data)

network_adhesion(.data)

network_diameter(.data)

network_length(.data)

network_independence(.data)

network_change(.data, object2)

network_stability(.data, object2)

network_reciprocity(.data, method = "default")

network_transitivity(.data)

network_equivalency(.data)

network_transitivity(.data)

network_congruency(.data)

network_richness(.data, attribute)

network_diversity(.data, attribute, clusters = NULL)

network_heterophily(.data, attribute)

network_assortativity(.data)

network_spatial(.data, attribute)

network_connectedness(.data)

network_efficiency(.data)

network_upperbound(.data)

network_nodes(.data)

network_ties(.data)

network_dims(.data)

network_dims.data.frame(.data)

network_dims.matrix(.data)

network_dims.igraph(.data)

network_dims.network(.data)

network_node_attributes(.data)

network_tie_attributes(.data)

network_degree(.data, normalized = TRUE, direction = c("all", "out", "in"))

network_outdegree(.data, normalized = TRUE)

network_indegree(.data, normalized = TRUE)

network_betweenness(
  .data,
  normalized = TRUE,
  direction = c("all", "out", "in")
)

network_closeness(.data, normalized = TRUE, direction = c("all", "out", "in"))

network_reach(.data, normalized = TRUE, k = 2)

network_harmonic(.data, normalized = TRUE, k = 2)

network_eigenvector(.data, normalized = TRUE)

network_core(.data, mark = NULL)

network_richclub(.data)

network_factions(.data, membership = NULL)

network_modularity(.data, membership = NULL, resolution = 1)

network_smallworld(.data, method = c("omega", "sigma", "SWI"), times = 100)

network_scalefree(.data)

network_balance(.data)

node_tie_census(.data)

node_triad_census(.data)

node_quad_census(.data)

node_path_census(.data)

network_dyad_census(.data)

network_triad_census(.data)

network_mixed_census(.data, object2)

node_brokerage_census(.data, membership, standardized = FALSE)

network_brokerage_census(.data, membership, standardized = FALSE)

node_brokering(.data, membership)

node_core(.data)

Functions

pkg_data(): Deprecated on 2024-06-17.

node_mode(): Deprecated on 2024-06-17.

autographr(): Deprecated on 2024-06-17.

autographs(): Deprecated on 2024-06-17.

autographd(): Deprecated on 2024-06-17.

node_optimal(): Deprecated on 2024-06-14.

node_kernighanlin(): Deprecated on 2024-06-14.

node_edge_betweenness(): Deprecated on 2024-06-14.

node_fast_greedy(): Deprecated on 2024-06-14.

node_leading_eigen(): Deprecated on 2024-06-14.

node_walktrap(): Deprecated on 2024-06-14.

node_infomap(): Deprecated on 2024-06-14.

node_spinglass(): Deprecated on 2024-06-14.

node_fluid(): Deprecated on 2024-06-14.

node_leiden(): Deprecated on 2024-06-14.

node_louvain(): Deprecated on 2024-06-14.

node_roulette(): Deprecated on 2024-06-14.

node_components(): Deprecated on 2024-06-14.

node_weak_components(): Deprecated on 2024-06-14.

node_strong_components(): Deprecated on 2024-06-14.

node_equivalence(): Deprecated on 2024-06-14.

node_structural_equivalence(): Deprecated on 2024-06-14.

node_regular_equivalence(): Deprecated on 2024-06-14.

node_automorphic_equivalence(): Deprecated on 2024-06-14.

network_transmissibility(): Deprecated on 2024-06-20.

network_infection_length(): Deprecated on 2024-06-20.

network_reproduction(): Deprecated on 2024-06-20.

network_immunity(): Deprecated on 2024-06-20.

network_hazard(): Deprecated on 2024-06-20.

network_density(): Deprecated on 2024-06-20.

network_components(): Deprecated on 2024-06-20.

network_cohesion(): Deprecated on 2024-06-20.

network_adhesion(): Deprecated on 2024-06-20.

network_diameter(): Deprecated on 2024-06-20.

network_length(): Deprecated on 2024-06-20.

network_independence(): Deprecated on 2024-06-20.

network_change(): Deprecated on 2024-06-20.

network_stability(): Deprecated on 2024-06-20.

network_reciprocity(): Deprecated on 2024-06-20.

network_transitivity(): Deprecated on 2024-06-20.

network_equivalency(): Deprecated on 2024-06-20.

network_transitivity(): Deprecated on 2024-06-20.

network_congruency(): Deprecated on 2024-06-20.

network_richness(): Deprecated on 2024-06-20.

network_diversity(): Deprecated on 2024-06-20.

network_heterophily(): Deprecated on 2024-06-20.

network_assortativity(): Deprecated on 2024-06-20.

network_spatial(): Deprecated on 2024-06-20.

network_connectedness(): Deprecated on 2024-06-20.

network_efficiency(): Deprecated on 2024-06-20.

network_upperbound(): Deprecated on 2024-06-20.

network_nodes(): Deprecated on 2024-06-20.

network_ties(): Deprecated on 2024-06-20.

network_dims(): Deprecated on 2024-06-20.

network_dims.data.frame(): Deprecated on 2024-06-20.

network_dims.matrix(): Deprecated on 2024-06-20.

network_dims.igraph(): Deprecated on 2024-06-20.

network_dims.network(): Deprecated on 2024-06-20.

network_node_attributes(): Deprecated on 2024-06-20.

network_tie_attributes(): Deprecated on 2024-06-20.

network_degree(): Deprecated on 2024-06-20.

network_outdegree(): Deprecated on 2024-06-20.

network_indegree(): Deprecated on 2024-06-20.

network_betweenness(): Deprecated on 2024-06-20.

network_closeness(): Deprecated on 2024-06-20.

network_reach(): Deprecated on 2024-06-20.

network_harmonic(): Deprecated on 2024-06-20.

network_eigenvector(): Deprecated on 2024-06-20.

network_core(): Deprecated on 2024-06-20.

network_richclub(): Deprecated on 2024-06-20.

network_factions(): Deprecated on 2024-06-20.

network_modularity(): Deprecated on 2024-06-20.

network_smallworld(): Deprecated on 2024-06-20.

network_scalefree(): Deprecated on 2024-06-20.

network_balance(): Deprecated on 2024-06-20.

node_tie_census(): Deprecated on 2024-06-20.

node_triad_census(): Deprecated on 2024-06-20.

node_quad_census(): Deprecated on 2024-06-20.

node_path_census(): Deprecated on 2024-06-20.

network_dyad_census(): Deprecated on 2024-06-20.

network_triad_census(): Deprecated on 2024-06-20.

network_mixed_census(): Deprecated on 2024-06-20.

node_brokerage_census(): Deprecated on 2024-06-20.

network_brokerage_census(): Deprecated on 2024-06-20.

node_brokering(): Deprecated on 2024-06-20.

node_core(): Deprecated on 2024-06-21.