Two-mode projection examples (Hollway 2021)

These networks are for demonstration purposes and do not describe any real world network. All examples contain named nodes. The networks are gathered together as a list and can be retrieved simply by plucking the desired network.

data(ison_laterals)

Format

#> $ison_bb
#> # A labelled, two-mode network of 10 nodes and 12 ties
#>
#> -- Nodes
#> # A tibble: 10 x 2
#>   name  type
#>   <chr> <lgl>
#> 1 A     FALSE
#> 2 B     FALSE
#> 3 C     FALSE
#> 4 D     FALSE
#> 5 U     TRUE
#> 6 V     TRUE
#> # i 4 more rows
#>
#> -- Ties
#> # A tibble: 12 x 2
#>    from    to
#>   <int> <int>
#> 1     1     5
#> 2     1     6
#> 3     2     5
#> 4     2     7
#> 5     2     8
#> 6     2     9
#> # i 6 more rows
#>
#>
#> $ison_bm
#> # A labelled, two-mode network of 8 nodes and 9 ties
#>
#> -- Nodes
#> # A tibble: 8 x 2
#>   name  type
#>   <chr> <lgl>
#> 1 A     FALSE
#> 2 B     FALSE
#> 3 C     FALSE
#> 4 D     FALSE
#> 5 U     TRUE
#> 6 V     TRUE
#> # i 2 more rows
#>
#> -- Ties
#> # A tibble: 9 x 2
#>    from    to
#>   <int> <int>
#> 1     1     5
#> 2     1     6
#> 3     2     5
#> 4     2     7
#> 5     2     8
#> 6     3     6
#> # i 3 more rows
#>
#>
#> $ison_mb
#> # A labelled, two-mode network of 8 nodes and 9 ties
#>
#> -- Nodes
#> # A tibble: 8 x 2
#>   name  type
#>   <chr> <lgl>
#> 1 A     FALSE
#> 2 B     FALSE
#> 3 C     FALSE
#> 4 D     FALSE
#> 5 M     TRUE
#> 6 X     TRUE
#> # i 2 more rows
#>
#> -- Ties
#> # A tibble: 9 x 2
#>    from    to
#>   <int> <int>
#> 1     1     5
#> 2     2     5
#> 3     2     6
#> 4     2     7
#> 5     3     5
#> 6     3     6
#> # i 3 more rows
#>
#>
#> $ison_mm
#> # A labelled, two-mode network of 6 nodes and 6 ties
#>
#> -- Nodes
#> # A tibble: 6 x 2
#>   name  type
#>   <chr> <lgl>
#> 1 A     FALSE
#> 2 B     FALSE
#> 3 C     FALSE
#> 4 D     FALSE
#> 5 M     TRUE
#> 6 N     TRUE
#>
#> -- Ties
#> # A tibble: 6 x 2
#>    from    to
#>   <int> <int>
#> 1     1     5
#> 2     2     5
#> 3     2     6
#> 4     3     5
#> 5     3     6
#> 6     4     6
#>