Membership of nodes in a diffusion

node_in_adopter() classifies membership of nodes into diffusion categories by where on the distribution of adopters they fell. Valente (1995) defines five memberships:

Early adopter: those with an adoption time less than the average adoption time minus one standard deviation of adoptions times

Early majority: those with an adoption time between the average adoption time and the average adoption time minus one standard deviation of adoptions times

Late majority: those with an adoption time between the average adoption time and the average adoption time plus one standard deviation of adoptions times

Laggard: those with an adoption time greater than the average adoption time plus one standard deviation of adoptions times

Non-adopter: those without an adoption time, i.e. never adopted

node_in_adopter(.data)

Arguments

.data

An object of a manynet-consistent class:

matrix (adjacency or incidence) from {base} R

edgelist, a data frame from {base} R or tibble from {tibble}

igraph, from the {igraph} package

network, from the {network} package

tbl_graph, from the {tidygraph} package

References

On adopter classes

Valente, Tom W. 1995. Network models of the diffusion of innovations (2nd ed.). Cresskill N.J.: Hampton Press.

See also

Examples

  smeg <- generate_smallworld(15, 0.025)
  smeg_diff <- play_diffusion(smeg, recovery = 0.2)
  # To classify nodes by their position in the adoption curve
  (adopts <- node_in_adopter(smeg_diff))
#> 4 groups
#>   V1     V2    V3    V4    V5    V6    V7    V8    V9    V10   V11   V12   V13  
#> 1 Early… Late… Earl… Late… Late… Late… Lagg… Late… Late… Late… Late… Earl… Earl…
#> # ... with 2 more values from this nodeset unprinted. Use `print(..., n = Inf)` to print all values.
  summary(adopts)
#> Class Early Adopter:  1, 3, 14, 15
#> Class Early Majority:  12, 13
#> Class Laggard:  7
#> Class Late Majority:  2, 4, 5, 6, 8, 9, 10, 11