Network permutation

to_permuted() permutes the network using a Fisher-Yates shuffle on both the rows and columns (for a one-mode network) or on each of the rows and columns (for a two-mode network).

to_permuted(.data, with_attr = TRUE)

Arguments

.data

An object of a {manynet}-consistent class:

adjacency or incidence matrix from {base} R

edgelist data.frame from {base} R or tbl/tbl_df from {tibble}

stocnet stocnet, from the {manynet} package

igraph igraph, from the {igraph} package

network network, from the {network} package

tidygraph tbl_graph, from the {tidygraph} package

with_attr

Logical whether any attributes of the object should be retained. By default TRUE.

Value

A tidygraph object modified as explained in the function description, details, or section.

See also