Motifs at the nodal level

These functions include ways to take a census of the positions of nodes in a network:

node_by_tie() returns a census of the ties in a network. For directed networks, out-ties and in-ties are bound together. for multiplex networks, the various types of ties are bound together.

node_by_triad() returns a census of the triad configurations nodes are embedded in.

node_by_quad() returns a census of nodes' positions in motifs of four nodes.

node_by_path() returns the shortest path lengths of each node to every other node in the network.

node_by_tie(.data)

node_by_triad(.data)

node_by_quad(.data)

node_by_path(.data)

Arguments

.data

An object of a manynet-consistent class:

matrix (adjacency or incidence) from {base} R

edgelist, a data frame from {base} R or tibble from {tibble}

igraph, from the {igraph} package

network, from the {network} package

tbl_graph, from the {tidygraph} package

Quad census

The quad census uses the {oaqc} package to do the heavy lifting of counting the number of each orbits. See vignette('oaqc'). However, our function relabels some of the motifs to avoid conflicts and improve some consistency with other census-labelling practices. The letter-number pairing of these labels indicate the number and configuration of ties. For now, we offer a rough translation:

migraph	Ortmann and Brandes

E4	co-K4

I40, I41	co-diamond

H4	co-C4

L42, L41, L40	co-paw

D42, D40	co-claw

U42, U41	P4

Y43, Y41	claw

P43, P42, P41	paw

04	C4

Z42, Z43	diamond

X4	K4

See also this list of graph classes.

References

Davis, James A., and Samuel Leinhardt. 1967. “The Structure of Positive Interpersonal Relations in Small Groups.” 55.

Ortmann, Mark, and Ulrik Brandes. 2017. “Efficient Orbit-Aware Triad and Quad Census in Directed and Undirected Graphs.” Applied Network Science 2(1):13. doi:10.1007/s41109-017-0027-2 .

Dijkstra, Edsger W. 1959. "A note on two problems in connexion with graphs". Numerische Mathematik 1, 269-71. doi:10.1007/BF01386390 .

Opsahl, Tore, Filip Agneessens, and John Skvoretz. 2010. "Node centrality in weighted networks: Generalizing degree and shortest paths". Social Networks 32(3): 245-51. doi:10.1016/j.socnet.2010.03.006 .

See also

Examples

task_eg <- to_named(to_uniplex(ison_algebra, "tasks"))
(tie_cen <- node_by_tie(task_eg))
#> # A tibble: 16 × 33
#>   names    fromAdalyn fromBernice fromCelia fromDianne fromEllen fromFranklin
#>   <chr>         <dbl>       <dbl>     <dbl>      <dbl>     <dbl>        <dbl>
#> 1 Adalyn          0             0      0             0      0.3          0   
#> 2 Bernice         0             0      0             0      0            0   
#> 3 Celia           0             0      0             0      0.15         0.75
#> 4 Dianne          0             0      0             0      0            0   
#> 5 Ellen           0.3           0      0.15          0      0            0.45
#> 6 Franklin        0             0      0.6           0      0.75         0   
#> # ℹ 10 more rows
#> # ℹ 26 more variables: fromGracie <dbl>, fromHeather <dbl>, fromIsabelle <dbl>,
#> #   fromJose <dbl>, fromKatelyn <dbl>, fromLizzie <dbl>, fromMarquita <dbl>,
#> #   fromNettie <dbl>, fromOlivia <dbl>, fromPiper <dbl>, toAdalyn <dbl>,
#> #   toBernice <dbl>, toCelia <dbl>, toDianne <dbl>, toEllen <dbl>,
#> #   toFranklin <dbl>, toGracie <dbl>, toHeather <dbl>, toIsabelle <dbl>,
#> #   toJose <dbl>, toKatelyn <dbl>, toLizzie <dbl>, toMarquita <dbl>, …
(triad_cen <- node_by_triad(task_eg))
#> # A tibble: 16 × 17
#>   names   X003  X012  X102 X021D X021U X021C X111D X111U X030T X030C  X201 X120D
#>   <chr>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 Adalyn    23    11    39     0     0     0     2     2     0     0    17     0
#> 2 Berni…    21     6    47     0     0     0     5     6     0     0    10     0
#> 3 Celia     34     3    46     0     0     0     2     1     0     0    13     0
#> 4 Dianne    52     6    28     0     0     0     0     0     0     0    18     0
#> 5 Ellen     17    11    37     1     0     1     6     7     0     0    15     1
#> 6 Frank…    23     8    40     1     0     0     3     3     0     0    18     1
#> # ℹ 10 more rows
#> # ℹ 4 more variables: X120U <dbl>, X120C <dbl>, X210 <dbl>, X300 <dbl>
node_by_quad(manynet::ison_southern_women)
#> # A tibble: 32 × 13
#>   names     E4   I41   I40    H4   L42   L41   L40   U42   U41   Y43   Y41   X04
#>   <chr>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 Evelyn  1463  1072   440   176   397   698   121   200    72    56   190    75
#> 2 Laura   1604  1008   541   180   323   665   155   134    79    35   168    68
#> 3 There…  1562   999   396   123   364   792    66   217    69    56   225    91
#> 4 Brenda  1624   990   524   176   320   688   152   134    76    35   170    71
#> 5 Charl…  2018   777   930   157   117   421   328    30    87     4    70    21
#> 6 Franc…  2138   664   898   114   102   506   240    36    98     4   130    30
#> # ℹ 26 more rows
node_by_path(manynet::ison_adolescents)
#> # A tibble: 8 × 9
#>   names Betty   Sue Alice  Jane  Dale   Pam Carol  Tina
#>   <chr> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 Betty     0     1     2     3     2     2     3     4
#> 2 Sue       1     0     1     2     1     1     2     3
#> 3 Alice     2     1     0     1     1     1     2     3
#> 4 Jane      3     2     1     0     1     2     3     4
#> 5 Dale      2     1     1     1     0     2     3     4
#> 6 Pam       2     1     1     2     2     0     1     2
#> # ℹ 2 more rows
node_by_path(manynet::ison_southern_women)
#> # A tibble: 32 × 33
#>   names  Evelyn Laura Theresa Brenda Charlotte Frances Eleanor Pearl  Ruth Verne
#>   <chr>   <dbl> <dbl>   <dbl>  <dbl>     <dbl>   <dbl>   <dbl> <dbl> <dbl> <dbl>
#> 1 Evelyn      0     2       2      2         2       2       2     2     2     2
#> 2 Laura       2     0       2      2         2       2       2     2     2     2
#> 3 There…      2     2       0      2         2       2       2     2     2     2
#> 4 Brenda      2     2       2      0         2       2       2     2     2     2
#> 5 Charl…      2     2       2      2         0       2       2     4     2     2
#> 6 Franc…      2     2       2      2         2       0       2     2     2     2
#> # ℹ 26 more rows
#> # ℹ 22 more variables: Myra <dbl>, Katherine <dbl>, Sylvia <dbl>, Nora <dbl>,
#> #   Helen <dbl>, Dorothy <dbl>, Olivia <dbl>, Flora <dbl>, E1 <dbl>, E2 <dbl>,
#> #   E3 <dbl>, E4 <dbl>, E5 <dbl>, E6 <dbl>, E7 <dbl>, E8 <dbl>, E9 <dbl>,
#> #   E10 <dbl>, E11 <dbl>, E12 <dbl>, E13 <dbl>, E14 <dbl>